swapped matching func
