+++ application/comparison_reporting/reporting.py
@@ -1,10 +1,8 @@
 from os import path
 import io
-from itertools import groupby
 from typing import List
 import csv
 from collections import defaultdict
-from datetime import datetime
 from pandas import DataFrame
 import pathlib
 import requests
@@ -13,13 +11,11 @@ from aws_lambda_powertools import Logger
 
 from common.nhs import NHSEntity
 from common.dos import DoSService, get_all_valid_dos_postcodes
-from common.opening_times import StandardOpeningTimes, OpenPeriod, SpecifiedOpeningTime, WEEKDAYS
+from common.opening_times import OpenPeriod, SpecifiedOpeningTime, WEEKDAYS
 
 logger = Logger(child=True)
 THIS_DIR = pathlib.Path(__file__).parent.resolve()
 OUTPUT_DIR = path.join(THIS_DIR, "out")
-DENTIST_DATA_FILE_URL = "https://assets.nhs.uk/data/foi/Dentists.csv"
-DENTIST_OPENING_TIMES_DATA_FILE_URL = "https://assets.nhs.uk/data/foi/DentistOpeningTimes.csv"
 
 
 def download_csv_as_dicts(url: str, delimiter: str = ",") -> List[dict]:
@@ -30,93 +26,17 @@ def download_csv_as_dicts(url: str, delimiter: str = ",") -> List[dict]:
             for row in csv.DictReader(csv_file_like_obj, skipinitialspace=True, delimiter=delimiter)]
 
 
-def get_dentists() -> List[NHSEntity]:
-    """Downloads the current NHS UK dentist data from https://www.nhs.uk/about-us/nhs-website-datasets/
-    and returns them as a list of NHSEntity objects"""
-
-    dentists_data = download_csv_as_dicts(DENTIST_DATA_FILE_URL, delimiter="¬")
-    dentists_open_times_data = download_csv_as_dicts(DENTIST_OPENING_TIMES_DATA_FILE_URL, delimiter="¬")
-
-    # Extract and sort opening times data items
-    std_open_data = defaultdict(list)
-    spec_open_data = defaultdict(list)
-    for item in dentists_open_times_data:
-        id = item["OrganisationId"]
-        opening_type = item["OpeningTimeType"]
-        if opening_type == "General":
-            std_open_data[id].append(item)
-        elif opening_type == "Additional":
-            spec_open_data[id].append(item)
-        else:
-            logger.warning(f"Unknown opening type '{opening_type}'.")
-
-    # Initialise dentists as NHS Entity objects
-    dentists = []
-    for entity_data in dentists_data:
-        id = entity_data.get("OrganisationID")
-
-        nhs_entity = NHSEntity({})
-        nhs_entity.odscode = entity_data.get("OrganisationCode")
-        nhs_entity.org_name = entity_data.get("OrganisationName")
-        nhs_entity.org_type = entity_data.get("OrganisationType")
-        nhs_entity.org_sub_type = entity_data.get("SubType")
-        nhs_entity.org_status = entity_data.get("OrganisationStatus")
-        nhs_entity.postcode = entity_data.get("Postcode")
-        nhs_entity.parent_org_name = entity_data.get("ParentName")
-        nhs_entity.city = entity_data.get("City")
-        nhs_entity.county = entity_data.get("County")
-        nhs_entity.address_lines = [
-            line
-            for line in [entity_data.get(x) for x in [f"Address{i}" for i in range(1, 5)] + ["City", "County"]]
-            if isinstance(line, str) and line.strip() != ""
-        ]
-        nhs_entity.phone = entity_data.get("Phone")
-        nhs_entity.website = entity_data.get("Website")
-        nhs_entity.standard_opening_times = StandardOpeningTimes()
-        nhs_entity.specified_opening_times = []
-
-        # Add std opening times to the dentist
-        for std_open_period in std_open_data.get(id, []):
-            is_open = std_open_period["IsOpen"].upper()
-            if is_open == "TRUE":
-                day = std_open_period["WeekDay"].lower()
-                open_period = OpenPeriod.from_string(std_open_period["Times"])
-                nhs_entity.standard_opening_times.add_open_period(open_period, day)
-            elif is_open == "FALSE":
-                logger.warning("A general weekday opening time IsOpen value is FALSE")
-
-        # Add spec opening times to the dentist
-        date_grouping = groupby(
-            spec_open_data.get(id, []),
-            lambda k: datetime.strptime(k["AdditonalOpeningDate"], "%b  %d  %Y").date()
-        )
-        for date, open_period_items in date_grouping:
-
-            is_open = True
-            open_periods = []
-            for item in list(open_period_items):
-                if item["IsOpen"].upper() == "TRUE":
-                    open_periods.append(OpenPeriod.from_string(item["Times"]))
-                else:
-                    is_open = False
-            nhs_entity.specified_opening_times.append(SpecifiedOpeningTime(open_periods, date, is_open))
-
-        dentists.append(nhs_entity)
-
-    return dentists
-
-
-def match_nhs_entiity_to_services(nhs_entities: List[NHSEntity], services: List[DoSService]):
+def match_nhs_entities_to_services(nhs_entities: List[NHSEntity], services: List[DoSService]):
     logger.info("Matching all NHS Entities to corresponding list of services.")
     servicelist_map = defaultdict(list)
     for nhs_entity in nhs_entities:
         for service in services:
-            if service.nhs_odscode_match(nhs_entity.odscode):
+            if nhs_entity.is_matching_dos_service(service):
                 servicelist_map[nhs_entity.odscode].append(service)
-    unmatched = len(nhs_entities) - len(servicelist_map)
+
     logger.info(
         f"{len(servicelist_map)}/{len(nhs_entities)} nhs entities matches with at least 1 service. "
-        f"{unmatched} not matched."
+        f"{len(nhs_entities) - len(servicelist_map)} not matched."
     )
     return servicelist_map
 
@@ -126,7 +46,7 @@ class Reporter:
     def __init__(self, nhs_entities: List[NHSEntity], dos_services: List[DoSService]):
         self.nhs_entities = nhs_entities
         self.dos_services = dos_services
-        self.entity_service_map = match_nhs_entiity_to_services(self.nhs_entities, self.dos_services)
+        self.entity_service_map = match_nhs_entities_to_services(self.nhs_entities, self.dos_services)
         self.valid_normalised_postcodes = None
 
     def create_postcode_comparison_report(self, filename: str):
